fixup! refactor(publish): read the publish queue from the DhtStore (Phase 2a)
